refactor(foundation): move ConditionEvaluator + HistoryManager to foundation (A4)
Phase 6 / Sub-phase A4: Both types are clean extracts.
- ConditionEvaluator (321 lines): zero crate deps (only std + serde_json + thiserror)
- HistoryManager (266 lines): only depends on StateMachine (already in foundation)

Moved to mockforge-foundation::state_machine::{condition_evaluator, history}.
Core re-exports preserved via thin shims.

Migrated mockforge-scenarios/src/state_machine.rs to import all four
(StateMachine, SubScenario, VisualLayout, ConditionEvaluator, HistoryManager)
from foundation instead of core.

Workspace builds clean.
